refactor: split monolithic lib into modules
